Some changes
Removed the cdn from tailwind and set it to a static tailwind file (This might make the app a little bit heavier, but it`s fine really.)

Trying to fix how the websocket connection stops and restarts.

Changed how the app retrieves the default theme too.

Changed how things are logged.