Small Update
So... I said that the YTP was going to be completed by the end of the week...

I was lying, sorry.

There’s quite a lot to do. I still need to reorganize the WebSocket connection state and send the signals for stopping YTP, Twitch’s WSS IRC, and the internal WebSocket using a single signal (I have no idea how to do that, lol).

There might be another week before we have a usable app.

There are a lot of bugs I haven’t discovered yet, simply because I can’t test this as thoroughly as I’d like to, since I don’t livestream that often and because I’m mostly working on this project while at work (don’t worry, I only code this after finishing all my daily tasks).

So that’s it. Expect another week of progress before a full release.