v0.1.2
---

**YouTube Integration Complete!**

At least, as of the last test on 15/09/2024, it worked perfectly without any errors—just a few odd behaviors, but that's on YouTube.

The app might be considered in a usable state, although I’m not 100% sure since I couldn’t test it as thoroughly as I’d like (you know, work, gym, life). I'll probably test it more in the future, but for now, it is what it is.

### Still To-Do:
1. Documentation page
2. Examples page
3. Profile page
4. Smooth out the resizing of the preview (it's really buggy at the moment)

This is more or less the priority list for what's next. That’s it for now!

---